new objects and fixes

+++ Sisao/utils.cpp
@@ -1,11 +1,11 @@
 #include "Utils.h"
 
 b2Vec2 to_box2d(glm::vec2 const& vec) {
-    return b2Vec2(vec.x, vec.y);
+    return b2Vec2((float)vec.x, (float)vec.y);
 }
 
 b2Vec2 to_box2d(glm::ivec2 const& vec) {
-    return b2Vec2(vec.x, vec.y);
+    return b2Vec2((float)vec.x, (float)vec.y);
 }
 
 glm::vec2 to_glm(b2Vec2 const& vec) {
